--- a/docs/en/guides/speed-estimation.md
+++ b/docs/en/guides/speed-estimation.md
@@ -44,36 +44,33 @@ keywords: Ultralytics YOLO11, speed estimation, object tracking, computer vision
 
         ```python
         import cv2
+
         from ultralytics import solutions
-        
+
         cap = cv2.VideoCapture("Path/to/video/file.mp4")
-        
+
         assert cap.isOpened(), "Error reading video file"
         w, h, fps = (int(cap.get(x)) for x in (cv2.CAP_PROP_FRAME_WIDTH, cv2.CAP_PROP_FRAME_HEIGHT, cv2.CAP_PROP_FPS))
-        
+
         video_writer = cv2.VideoWriter("speed_management.avi", cv2.VideoWriter_fourcc(*"mp4v"), fps, (w, h))
-        
+
         speed_region = [(20, 400), (1080, 404), (1080, 360), (20, 360)]
-        
-        speed = solutions.SpeedEstimator(
-            model="yolo11n.pt",
-            region=speed_region,
-            show=True
-        )
-        
+
+        speed = solutions.SpeedEstimator(model="yolo11n.pt", region=speed_region, show=True)
+
         while cap.isOpened():
             success, im0 = cap.read()
-        
+
             if success:
                 out = speed.estimate_speed(im0)
                 video_writer.write(im0)
                 if cv2.waitKey(1) & 0xFF == ord("q"):
                     break
                 continue
-        
+
             print("Video frame is empty or video processing has been successfully completed.")
             break
-        
+
         cap.release()
         cv2.destroyAllWindows()
         ```
@@ -84,12 +81,12 @@ keywords: Ultralytics YOLO11, speed estimation, object tracking, computer vision
 
 ### Arguments `SpeedEstimator`
 
-| Name         | Type   | Default                    | Description                                                            |
-| ------------ | ------ | -------------------------- | ---------------------------------------------------------------------- |
-| `model`      | `str`  | `None`                     | Path to Ultralytics YOLO Model File                                    |
-| `region`     | `list` | `[(20, 400), (1260, 400)]` | List of points defining the counting region.                           |
-| `line_width` | `int`  | `2`                        | Line thickness for bounding boxes.                                     |
-| `show`       | `bool` | `False`                    | Flag to control whether to display the video stream.                   |
+| Name         | Type   | Default                    | Description                                          |
+| ------------ | ------ | -------------------------- | ---------------------------------------------------- |
+| `model`      | `str`  | `None`                     | Path to Ultralytics YOLO Model File                  |
+| `region`     | `list` | `[(20, 400), (1260, 400)]` | List of points defining the counting region.         |
+| `line_width` | `int`  | `2`                        | Line thickness for bounding boxes.                   |
+| `show`       | `bool` | `False`                    | Flag to control whether to display the video stream. |
 
 ### Arguments `model.track`
 
@@ -106,7 +103,7 @@ Estimating object speed with Ultralytics YOLO11 involves combining [object detec
 ```python
 import cv2
 
-from ultralytics import YOLO, solutions
+from ultralytics import solutions
 
 cap = cv2.VideoCapture("path/to/video/file.mp4")
 w, h, fps = (int(cap.get(x)) for x in (cv2.CAP_PROP_FRAME_WIDTH, cv2.CAP_PROP_FRAME_HEIGHT, cv2.CAP_PROP_FPS))