@glenn-jocher This PR will only update the `speed-estimation` solution.

### Main changes
- The Model class will be instantiated within the solutions, so users only need to provide the model file path.
- Object tracking will now be handled within the solutions, instead of being managed by an external video processing loop.

### New Code
```python
import cv2
from ultralytics import solutions

cap = cv2.VideoCapture("Path/to/video/file.mp4")

assert cap.isOpened(), "Error reading video file"
w, h, fps = (int(cap.get(x)) for x in (cv2.CAP_PROP_FRAME_WIDTH, cv2.CAP_PROP_FRAME_HEIGHT, cv2.CAP_PROP_FPS))

video_writer = cv2.VideoWriter("speed_estimation.avi", cv2.VideoWriter_fourcc(*"mp4v"), fps, (w, h))

speed = solutions.SpeedEstimator(
    model="yolo11n.pt",
    region= [(20, 400), (1080, 404), (1080, 360), (20, 360)],
    show=True
)

while cap.isOpened():
    success, im0 = cap.read()
    if success:
        out = speed.estimate_speed(im0)
        video_writer.write(im0)
        if cv2.waitKey(1) & 0xFF == ord("q"):
            break
        continue
    print("Video frame is empty or video processing has been successfully completed.")
    break

cap.release()
cv2.destroyAllWindows()
```

### Output Sample

```CLI
Ultralytics Solutions: ✅ {'model': 'yolo11n.pt', 'region': [(20, 400), (1080, 404), (1080, 360), (20, 360)], 'line_width': 2, 'show': True, 'show_in': True, 'show_out': True, 'classes': None, 'up_angle': 145.0, 'down_angle': 90, 'kpts': [6, 8, 10], 'colormap': None}

0: 384x640 7 persons, 203.6ms
Speed: 15.6ms preprocess, 203.6ms inference, 0.0ms postprocess per image at shape (1, 3, 384, 640)
```

## 🛠️ PR Summary

<sub>Made with ❤️ by [Ultralytics Actions](https://github.com/ultralytics/actions)<sub>

### 🌟 Summary
Improved the SpeedEstimator class for more efficient speed tracking and estimation in video streams.

### 📊 Key Changes
- Revised the initialization and tracking process in `SpeedEstimator`.
- Simplified code by removing unnecessary imports and variables.
- Updated speed estimation algorithm for better accuracy.
- Changed parameters for drawing regions and handling video processing.

### 🎯 Purpose & Impact
- 🏎️ **Better Performance**: Streamlines speed estimation for faster and more accurate results.
- 🖥️ **Code Simplification**: Makes it easier for developers to integrate and maintain.
- 🎥 **User Experience**: Enhancements improve video processing reliability and feedback.