@UltralyticsAssistant 2024-10-09
👋 Hello @RizwanMunawar, thank you for submitting an `ultralytics/ultralytics` 🚀 PR! To ensure a seamless integration of your work, please review the following checklist:

- ✅ **Define a Purpose**: Clearly explain the intention behind your updates to the `speed-estimation` solution. Be sure to reference any [relevant issues](https://github.com/ultralytics/ultralytics/issues) if applicable. Ensure your commit messages are succinct and follow project conventions.
  
- ✅ **Synchronize with Source**: Make sure your PR is in sync with the `main` branch of `ultralytics/ultralytics`. Update if necessary by clicking the 'Update branch' button or running `git pull` and `git merge main` locally.
  
- ✅ **Ensure CI Checks Pass**: Verify that all Ultralytics [Continuous Integration (CI)](https://docs.ultralytics.com/help/CI/) checks are successful. Address any failing checks promptly.
  
- ✅ **Update Documentation**: Update the relevant [documentation](https://docs.ultralytics.com) to reflect new or modified features in your PR.
  
- ✅ **Add Tests**: If applicable, include or update tests to ensure your changes are covered, and verify all tests pass.
  
- ✅ **Sign the CLA**: Ensure you have signed our [Contributor License Agreement](https://docs.ultralytics.com/help/CLA/). Add a message with "I have read the CLA Document and I sign the CLA" if this is your first PR.
  
- ✅ **Minimize Changes**: Limit your modifications to those necessary for your feature or bug fix. _"The closer to the source, the less wastage there is."_ — Bruce Lee

If your PR is related to a bug, please include a minimum reproducible example (MRE) if you haven't already. For more guidance, refer to our [Contributing Guide](https://docs.ultralytics.com/help/contributing). Feel free to leave a comment if you have any questions. Thank you for contributing to Ultralytics! 🚀

This is an automated response, and an Ultralytics engineer will review your submission soon. 😊